taskコードの調整変更
