+++ backer/godotask/model/Task.go
@@ -14,3 +14,7 @@ type Task struct {
 	CreatedAt   time.Time  `json:"created_at"`
 	UpdatedAt   time.Time  `json:"updated_at"`
 }
+
+func (Task) TableName() string {
+    return "task"
+}
